requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Pillow==10.0.0
PyPDF2==3.0.1
ebooklib==0.18
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
requests==2.31.0
beautifulsoup4==4.12.2
